requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0